requirements.txt
python-telegram-bot>=20.0
pyyaml==6.0
aiohttp==3.8.4
cryptography==40.0.2
qrcode==7.4.2
pillow==9.5.0
asyncio==3.4.3
python-dotenv==1.0.0 